Prototype, mark functions as __smallc
Change section name to _splib2, use bss/data/rodata as necessary

Point heap/ to adt_Heap in the library

+++ cwrapper/sp_inp.c
@@ -7,7 +7,7 @@
 #define _SPLIB
 #include "spritepack.h"
 
-uchar sp_inp(uint port)
+uint8_t sp_inp(uint16_t port)
 {
 #asm
    ld hl,2
